Python code
The final network that has been chosen based on their accuracy to classify these images is as below: 
For the first layer, flatten and dense layer are implemented. For the hidden layer Relu as activation function, then dropout and dense layer are used.  A dropout probability of 0.5 is applied during the training phase between the fully connected layers. Sigmoid activation function is applied at the output layer. Sigmoid function will quash the output values to a range between 0 and 1 making it suitable for binary image classification problems like this one. The model uses rmsprop as an optimiser to minimise binary cross entropy.